+++ sdk/resourcemanager/datafactory/armdatafactory/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Release History
 
+## 4.1.0 (2023-12-15)
+### Features Added
+
+- New function `*WarehouseLinkedService.GetLinkedService() *LinkedService`
+- New function `*WarehouseSink.GetCopySink() *CopySink`
+- New function `*WarehouseSource.GetCopySource() *CopySource`
+- New function `*WarehouseSource.GetTabularSource() *TabularSource`
+- New function `*WarehouseTableDataset.GetDataset() *Dataset`
+- New struct `WarehouseLinkedService`
+- New struct `WarehouseLinkedServiceTypeProperties`
+- New struct `WarehouseSink`
+- New struct `WarehouseSource`
+- New struct `WarehouseTableDataset`
+- New struct `WarehouseTableDatasetTypeProperties`
+- New field `Metadata` in struct `AzureBlobFSWriteSettings`
+- New field `Metadata` in struct `AzureBlobStorageWriteSettings`
+- New field `Metadata` in struct `AzureDataLakeStoreWriteSettings`
+- New field `Metadata` in struct `AzureFileStorageWriteSettings`
+- New field `Metadata` in struct `FileServerWriteSettings`
+- New field `Metadata` in struct `LakeHouseWriteSettings`
+- New field `Metadata` in struct `SftpWriteSettings`
+- New field `Metadata` in struct `StoreWriteSettings`
+- New field `HTTPRequestTimeout`, `TurnOffAsync` in struct `WebActivityTypeProperties`
+
+
 ## 4.0.0 (2023-12-22)
 ### Breaking Changes